1: Refactor folder structure. Add CMake and auto-build script.
