+++ script/build
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+# Create build directory if it doesn't exist
+mkdir -p build
+
+# Navigate to build directory
+cd build
+
+# Run cmake
+cmake ..
+
+# Run make
+make
